+++ components/search/src/filter.rs
@@ -22,7 +22,7 @@ impl From<JSONEngineUrl> for SearchEngineUrl {
             method: url.method.unwrap_or_default().as_str().to_string(),
             params: url.params.unwrap_or_default(),
             search_term_param_name: url.search_term_param_name,
-            display_name: url.display_name,
+            display_name_map: url.display_name_map,
         }
     }
 }
@@ -41,8 +41,8 @@ impl SearchEngineUrl {
         if let Some(search_term_param_name) = &preferred.search_term_param_name {
             self.search_term_param_name = Some(search_term_param_name.clone());
         }
-        if let Some(display_name) = &preferred.display_name {
-            self.display_name = Some(display_name.clone());
+        if let Some(display_name_map) = &preferred.display_name_map {
+            self.display_name_map = Some(display_name_map.clone());
         }
     }
 }
@@ -474,7 +474,7 @@ fn find_engine_id_with_match(
 
 #[cfg(test)]
 mod tests {
-    use std::vec;
+    use std::{collections::HashMap, vec};
 
     use super::*;
     use crate::*;
@@ -501,7 +501,7 @@ mod tests {
                     method: None,
                     params: None,
                     search_term_param_name: None,
-                    display_name: None,
+                    display_name_map: None,
                 }),
                 ..Default::default()
             },
@@ -547,7 +547,7 @@ mod tests {
                         method: None,
                         params: None,
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     suggestions: None,
                     trending: None,
@@ -589,7 +589,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     },
                     suggestions: None,
                     trending: None,
@@ -626,7 +626,7 @@ mod tests {
                     },
                 ]),
                 search_term_param_name: Some("baz".to_string()),
-                display_name: None,
+                display_name_map: None,
             }),
             suggestions: Some(JSONEngineUrl {
                 base: Some("https://example.com/suggestions".to_string()),
@@ -638,7 +638,7 @@ mod tests {
                     experiment_config: Some("suggest-experiment-value".to_string()),
                 }]),
                 search_term_param_name: Some("suggest".to_string()),
-                display_name: None,
+                display_name_map: None,
             }),
             trending: Some(JSONEngineUrl {
                 base: Some("https://example.com/trending".to_string()),
@@ -650,7 +650,7 @@ mod tests {
                     experiment_config: None,
                 }]),
                 search_term_param_name: None,
-                display_name: None,
+                display_name_map: None,
             }),
             search_form: Some(JSONEngineUrl {
                 base: Some("https://example.com/search_form".to_string()),
@@ -662,7 +662,7 @@ mod tests {
                     experiment_config: None,
                 }]),
                 search_term_param_name: None,
-                display_name: None,
+                display_name_map: None,
             }),
             visual_search: Some(JSONEngineUrl {
                 base: Some("https://example.com/visual_search".to_string()),
@@ -674,7 +674,10 @@ mod tests {
                     experiment_config: None,
                 }]),
                 search_term_param_name: Some("url".to_string()),
-                display_name: Some("Visual Search".to_string()),
+                display_name_map: Some(HashMap::from([(
+                    "default".to_string(),
+                    "Visual Search".to_string(),
+                )])),
             }),
         },
     });
@@ -731,7 +734,7 @@ mod tests {
                             },
                         ],
                         search_term_param_name: Some("baz".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions".to_string(),
@@ -743,7 +746,7 @@ mod tests {
                             experiment_config: Some("suggest-experiment-value".to_string()),
                         }],
                         search_term_param_name: Some("suggest".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending".to_string(),
@@ -755,7 +758,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search_form".to_string(),
@@ -767,7 +770,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(SearchEngineUrl {
                         base: "https://example.com/visual_search".to_string(),
@@ -779,7 +782,10 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("url".to_string()),
-                        display_name: Some("Visual Search".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search".to_string()
+                        )])),
                     }),
                 },
                 click_url: None
@@ -812,7 +818,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("ship".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-variant".to_string()),
@@ -824,7 +830,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("variant".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-variant".to_string()),
@@ -836,7 +842,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("trend".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search_form".to_string()),
@@ -848,7 +854,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(JSONEngineUrl {
                         base: Some("https://example.com/visual-search-variant".to_string()),
@@ -860,7 +866,10 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("url_variant".to_string()),
-                        display_name: Some("Visual Search Variant".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search Variant".to_string(),
+                        )])),
                     }),
                 }),
                 sub_variants: vec![],
@@ -892,7 +901,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("ship".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions-variant".to_string(),
@@ -904,7 +913,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("variant".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending-variant".to_string(),
@@ -916,7 +925,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("trend".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search_form".to_string(),
@@ -928,7 +937,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(SearchEngineUrl {
                         base: "https://example.com/visual-search-variant".to_string(),
@@ -940,7 +949,10 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("url_variant".to_string()),
-                        display_name: Some("Visual Search Variant".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search Variant".to_string()
+                        )])),
                     }),
                 },
                 click_url: None
@@ -973,7 +985,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("ship".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-variant".to_string()),
@@ -985,7 +997,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("variant".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-variant".to_string()),
@@ -997,7 +1009,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("trend".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search-form-variant".to_string()),
@@ -1009,7 +1021,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(JSONEngineUrl {
                         base: Some("https://example.com/visual-search-variant".to_string()),
@@ -1021,7 +1033,10 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("url_variant".to_string()),
-                        display_name: Some("Visual Search Variant".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search Variant".to_string(),
+                        )])),
                     }),
                 }),
                 // This would be the list of sub-variants for this part of the
@@ -1050,7 +1065,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("shuttle".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     suggestions: Some(JSONEngineUrl {
                         base: Some("https://example.com/suggestions-subvariant".to_string()),
@@ -1062,7 +1077,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("subvariant".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(JSONEngineUrl {
                         base: Some("https://example.com/trending-subvariant".to_string()),
@@ -1074,7 +1089,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("subtrend".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(JSONEngineUrl {
                         base: Some("https://example.com/search-form-subvariant".to_string()),
@@ -1086,7 +1101,7 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(JSONEngineUrl {
                         base: Some("https://example.com/visual-search-subvariant".to_string()),
@@ -1098,7 +1113,10 @@ mod tests {
                             experiment_config: None,
                         }]),
                         search_term_param_name: Some("url_subvariant".to_string()),
-                        display_name: Some("Visual Search Subvariant".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search Subvariant".to_string(),
+                        )])),
                     }),
                 }),
                 sub_variants: vec![],
@@ -1129,7 +1147,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("shuttle".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     },
                     suggestions: Some(SearchEngineUrl {
                         base: "https://example.com/suggestions-subvariant".to_string(),
@@ -1141,7 +1159,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("subvariant".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     trending: Some(SearchEngineUrl {
                         base: "https://example.com/trending-subvariant".to_string(),
@@ -1153,7 +1171,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("subtrend".to_string()),
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     search_form: Some(SearchEngineUrl {
                         base: "https://example.com/search-form-subvariant".to_string(),
@@ -1165,7 +1183,7 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     }),
                     visual_search: Some(SearchEngineUrl {
                         base: "https://example.com/visual-search-subvariant".to_string(),
@@ -1177,7 +1195,10 @@ mod tests {
                             experiment_config: None,
                         }],
                         search_term_param_name: Some("url_subvariant".to_string()),
-                        display_name: Some("Visual Search Subvariant".to_string()),
+                        display_name_map: Some(HashMap::from([(
+                            "default".to_string(),
+                            "Visual Search Subvariant".to_string()
+                        )])),
                     }),
                 },
                 click_url: None
@@ -1196,7 +1217,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     },
                     ..Default::default()
                 },
@@ -1211,7 +1232,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     },
                     ..Default::default()
                 },
@@ -1226,7 +1247,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     },
                     ..Default::default()
                 },
@@ -1241,7 +1262,7 @@ mod tests {
                         method: "GET".to_string(),
                         params: Vec::new(),
                         search_term_param_name: None,
-                        display_name: None,
+                        display_name_map: None,
                     },
                     ..Default::default()
                 },