+++ CHANGELOG.md
@@ -23,8 +23,8 @@
 
 - `SearchEngineUrls` now has an optional `visual_search` field, supporting
   visual search endpoints in engine configs.
-- `SearchEngineUrl` now has an optional `display_name` field, which is useful if
-  a URL corresponds to a brand name distinct from the engine's brand name.
+- `SearchEngineUrl` now has an optional `display_name_map` field, which is useful
+  if a URL corresponds to a brand name distinct from the engine's brand name.
 
 [Full Changelog](https://github.com/mozilla/application-services/compare/v141.0...v142.0)